+++ src/plugins/intel_npu/src/utils/src/CMakeLists.txt
@@ -7,4 +7,5 @@ add_subdirectory(logger)
 if(ENABLE_NPU_PLUGIN_ENGINE)
     add_subdirectory(vcl)
     add_subdirectory(zero)
+    add_subdirectory(vm)
 endif()
